I'm having a hard time getting past an error: The projects in the reactor contain a cyclic reference: Edge between 'Vertex{label='com.tommontom.pdfsplitter:pdf-splitter-swing:0.0.1-SNAPSHOT'}' and 'Vertex{label='com.tommontom.pdfsplitter:pdf-splitter-core:0.0.1-SNAPSHOT'}' introduces to cycle in the graph com.tommontom.pdfsplitter:pdf-splitter-core:0.0.1-SNAPSHOT --> com.tommontom.pdfsplitter:pdf-splitter-swing:0.0.1-SNAPSHOT --> com.tommontom.pdfsplitter:pdf-splitter-core:0.0.1-SNAPSHOT -> [Help 1]
To see the full stack trace of the errors, re-run Maven with the -e switch.
Re-run Maven using the -X switch to enable full debug logging.

For more information about the errors and possible solutions, please read the following articles:
[Help 1] http://cwiki.apache.org/confluence/display/MAVEN/ProjectCycleException

This is after trying to build the project. It has to do with using the core as a dependency in the window main as well as windowmain as a dependency in the core.